WIP: Data filtering Another pass, this time currently removing the use of infinispan and just using a listener approach.  Also, consolidates the filter handling into a Singleton. And added avail data id filtering.
This all works fine where everything is co-located. What we may want to
do is keep it like this but also have the "cachemanager" drop info in the
cache in order to notify other servers of updates.

Also, currently the event info is very coarse, we could make things more
granular and/or provide more contextual data, perhaps even the affected
ids, and therefore avoid total refreshes of the active id lists.