HWKALERTS-28, prefer path params
The triggerId is required in and the target of the path so I think we
should use it.  Ensure the Condition is created using the 'triggerId' path
param.

Moreover, apply this semantic consistently, ensuring that path params are
assumed correct, and apply them to, or use them to verify, the ApiParam
structures.