HWKALERTS-101 Fix multi-tenancy issues
All tests/itests are passing. A little more soak testing but this should be
close to a PR...

MultiTenancy issues extended into condition and dampening handling because
of they referenced only triggerId, as opposed to the unique combination
of tenantId+triggerId. The rulebase has been updated to use tenantId.
Unfortunately, proper handling required updating of Condition and
Dampening constructors to include a tenantId parameter, which had a
"ripple effect", touchig a lot of files.

REST clients do not need to provide a tenantId because one is automatically
assigned. There are a few constructors in place for use by REST client code,
sothers can be used by supplying an "" tenantId.

We may want to discuss whether we need more robust support for the rest
clients.
others are available
constructors is currently