Add displayString support to Condition so that REST (#414)
clients (like the UI) can easily access a canonical display
string describing the Condition.  This is somewhat analogous
to the recent work to provide the same thing for ConditionEval
and the canonical display strings are similar for consistency.

Note, also for consistency, the poorly named ConditionEval.getLog()
is now ConditionEval.displayString(), to match Condition.displayString().